Set dep from glance cron jobs to api package
Currently there is no dependency between the
glance cron jobs and the glance api package.

This dependency needs to exist b/c the cron jobs
cannot even be created by Puppet until the
glance user exists on the system (which in
the case of package installation is installed
via the installation of the glance package)

This patch adds the dependency.

Change-Id: Ib0a42cf171d4fcabaab616c880aad2d0c4db4424