Fix catalog compilation when not configuring endpoint
Previously when you set $configure_endpoint to false and did not have a
keystone_endpoint resource with the same name as is defined by
glance::keystone::auth, the catalog would not be compiled. This was
because a relationship was being established where one resource in the
relationship didn't exist.

This changes this so that the relationship involving the Keystone
endpoint is only defined if $configure_endpoint is set to true.

Also, the test for when $configure_endpoint is false was set to check
for an endpoint that doesn't get created even if $configure_endpoint is
true. Fixed so that it checks for an endpoint that has the correct name
(the same as when using the default values to call the class).

Change-Id: I6631a656a888d0df7ab173705bea0598bd2b2f58
Closes-Bug: 1368686