Creation of glance::db::sync
In order to standardize the way dbsync are run across our modules,
we create a new class glance::db::sync.
This class will be included if sync_db is enabled.

By making this transition the glance::db::sync class
can be returned by the ENC.

A use case would be in an highly available environment, with 3 galera
nodes, include glance::registry on every node with sync_db set to false
and have the ENC return glance::db::sync just for one node.

Change-Id: I16c8bc411bd34f720513a5d4c94e82b07105af20