Make sure Glance_image is executed after Keystone_endpoint
When running Keystone in WSGI, Glance_image is not run in the right
moment.
This patch aims to make sure Glance_image is run after creating Glance
endpoint Keystone resource by using the Puppet chain arrow.
Keystone_endpoint is actually the latest resource created by Keystone
providers to make Glance working.
If Glance_image resource is in the catalog, it will be executed when
keystone & glance are actually ready and not before.

Closes-Bug: #1488277
Change-Id: I97d1a77c5f4a67914738514d773ecbd4bfb2196c